// src/api/v1/categories/interfaces.ts
export interface ICategory {
  id: number;
  title: string;
  article: string;
  article_city: string;
  article_state: string;
  meta_title: string;
  meta_title_city: string;
  meta_title_state: string;
  meta_description: string;
  meta_description_city: string;
  meta_description_state: string;
}